Enable JS in tsconfig and migrate several modules to TS
