--- src/shared/hooks/useProjectId.js
@@ -1,2 +1,0 @@
-// src/shared/hooks/useProjectId.js
-export { useProjectId } from '@/shared/store/authStore';
