+++ src/shared/hooks/useProjectId.ts
@@ -1,0 +1,3 @@
+// src/shared/hooks/useProjectId.ts
+export { useProjectId } from '@/shared/store/authStore';
+
