+++ tsconfig.json
@@ -16,7 +16,8 @@
       "@shared/*": ["shared/*"]
     },
     "esModuleInterop": true,
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "allowJs": true
   },
   "include": ["src"]
 }
